# ENH: spatial: faster 2D Voronoi and Convex Hull plotting