--- a/scipy/spatial/_plotutils.py
+++ b/scipy/spatial/_plotutils.py
@@ -9,6 +9,7 @@ __all__ = ['delaunay_plot_2d', 'convex_hull_plot_2d', 'voronoi_plot_2d']
 @_decorator
 def _held_figure(func, obj, ax=None, **kw):
     import matplotlib.pyplot as plt
+    from matplotlib.collections import LineCollection
 
     if ax is None:
         fig = plt.figure()
@@ -98,9 +99,12 @@ def convex_hull_plot_2d(hull, ax=None):
         raise ValueError("Convex hull is not 2-D")
 
     ax.plot(hull.points[:,0], hull.points[:,1], 'o')
+    line_segments = []
     for simplex in hull.simplices:
-        ax.plot(hull.points[simplex,0], hull.points[simplex,1], 'k-')
-
+        line_segments.append([(x, y) for x, y in hull.points[simplex]])
+    ax.add_collection(LineCollection(line_segments,
+                                     colors = 'k',
+                                     linestyle = 'solid'))
     _adjust_bounds(ax, hull.points)
 
     return ax.figure
@@ -138,13 +142,18 @@ def voronoi_plot_2d(vor, ax=None):
     ax.plot(vor.points[:,0], vor.points[:,1], '.')
     ax.plot(vor.vertices[:,0], vor.vertices[:,1], 'o')
 
+    line_segments = []
     for simplex in vor.ridge_vertices:
         simplex = np.asarray(simplex)
         if np.all(simplex >= 0):
-            ax.plot(vor.vertices[simplex,0], vor.vertices[simplex,1], 'k-')
+            line_segments.append([(x, y) for x, y in vor.vertices[simplex]])
 
+    ax.add_collection(LineCollection(line_segments,
+                                     colors = 'k',
+                                     linestyle = 'solid'))
     ptp_bound = vor.points.ptp(axis=0)
 
+    line_segments = []
     center = vor.points.mean(axis=0)
     for pointidx, simplex in zip(vor.ridge_points, vor.ridge_vertices):
         simplex = np.asarray(simplex)
@@ -159,9 +168,12 @@ def voronoi_plot_2d(vor, ax=None):
             direction = np.sign(np.dot(midpoint - center, n)) * n
             far_point = vor.vertices[i] + direction * ptp_bound.max()
 
-            ax.plot([vor.vertices[i,0], far_point[0]],
-                    [vor.vertices[i,1], far_point[1]], 'k--')
+            line_segments.append([(vor.vertices[i, 0], vor.vertices[i, 1]),
+                                  (far_point[0], far_point[1])])
 
+    ax.add_collection(LineCollection(line_segments,
+                                     colors = 'k',
+                                     linestyle = 'dashed'))
     _adjust_bounds(ax, vor.points)
 
     return ax.figure