This pull request modifies ``voronoi_plot_2d`` and ``convex_hull_plot_2d`` to use ``matplotlib.collections.LineCollection`` objects.  This removes the need for repeatedly calling ``ax.plot`` on each individual line segment and speeds up plotting.  This improvement is especially dramatic for the Voronoi case (approximately 13 times faster for me in the example below!).

On my PC, for a set of 2000 randomly generated 2D points I got the following speedups
**voronoi_plot_2d** :        previously 8.45 s -> now 0.63 s
**convex_hull_plot_2d**:|  previously 111 ms -> now 83 ms

I also added an optional kwarg to ``voronoi_plot_2d`` that can be used to avoid adding the vertices to the plot.  To my eye this looks cleaner, especially when there are a large number of points, but I have left the old behavior as the default.